(PC-7905): Move VenueProvider to models file
