+++ src/pcapi/admin/install.py
@@ -22,6 +22,7 @@ from pcapi.admin.custom_views.user_offerer_view import UserOffererView
 from pcapi.admin.custom_views.venue_provider_view import VenueProviderView
 from pcapi.admin.custom_views.venue_view import VenueView
 from pcapi.core.offerers.models import Offerer
+from pcapi.core.providers.models import VenueProvider
 from pcapi.core.users.models import User
 from pcapi.models import AllocinePivot
 from pcapi.models import ApiKey
@@ -31,7 +32,6 @@ from pcapi.models import Feature
 from pcapi.models import Offer
 from pcapi.models import UserOfferer
 from pcapi.models import Venue
-from pcapi.models import VenueProvider
 
 
 class Category(Enum):
