+++ src/pcapi/local_providers/provider_api/provider_api_stocks.py
@@ -3,13 +3,12 @@ import logging
 from sqlalchemy.orm.util import aliased
 
 from pcapi.core.providers.models import Provider
-from pcapi.core.providers.venue_provider import VenueProvider
+from pcapi.core.providers.models import VenueProvider
+from . import synchronize_provider_api
 from pcapi.models import db
 
 
 logger = logging.getLogger(__name__)
-
-from . import synchronize_provider_api
 
 
 def synchronize_stocks() -> None:
