+++ src/pcapi/local_providers/provider_api/synchronize_provider_api.py
@@ -14,7 +14,7 @@ from pcapi.core.offers.repository import get_offers_map_by_id_at_providers
 from pcapi.core.offers.repository import get_products_map_by_id_at_providers
 from pcapi.core.offers.repository import get_stocks_by_id_at_providers
 from pcapi.core.providers.models import Provider
-from pcapi.core.providers.venue_provider import VenueProvider
+from pcapi.core.providers.models import VenueProvider
 from pcapi.infrastructure.repository.stock_provider.provider_api import ProviderAPI
 from pcapi.models import Product
 from pcapi.models import Venue
