+++ tests/local_providers/provider_test_utils.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 
+from pcapi.core.providers.models import VenueProvider
 from pcapi.local_providers.local_provider import LocalProvider
 from pcapi.models import ThingType
-from pcapi.models import VenueProvider
 from pcapi.models.db import Model
 import pcapi.sandboxes
