+++ tests/models/allocine_venue_provider_test.py
@@ -1,5 +1,6 @@
 import pytest
 
+from pcapi.core.providers.models import VenueProvider
 from pcapi.model_creators.generic_creators import create_allocine_venue_provider
 from pcapi.model_creators.generic_creators import create_offerer
 from pcapi.model_creators.generic_creators import create_provider
@@ -7,7 +8,6 @@ from pcapi.model_creators.generic_creators import create_venue
 from pcapi.model_creators.generic_creators import create_venue_provider
 from pcapi.model_creators.provider_creators import activate_provider
 from pcapi.models import AllocineVenueProvider
-from pcapi.models import VenueProvider
 from pcapi.repository import repository
