+++ tests/models/venue_provider_test.py
@@ -1,5 +1,6 @@
 import pytest
 
+from pcapi.core.providers.models import VenueProvider
 from pcapi.model_creators.generic_creators import create_offerer
 from pcapi.model_creators.generic_creators import create_provider
 from pcapi.model_creators.generic_creators import create_venue
@@ -8,7 +9,6 @@ from pcapi.model_creators.provider_creators import activate_provider
 from pcapi.model_creators.specific_creators import create_offer_with_event_product
 from pcapi.model_creators.specific_creators import create_offer_with_thing_product
 from pcapi.models import ApiErrors
-from pcapi.models import VenueProvider
 from pcapi.repository import repository
